+++ dags/web_preaggregated_daily.py
@@ -3,7 +3,7 @@ from collections.abc import Callable
 import os
 
 import dagster
-from dagster import DailyPartitionsDefinition, BackfillPolicy, asset_check, AssetCheckResult, MetadataValue
+from dagster import DailyPartitionsDefinition, BackfillPolicy
 import structlog
 from dags.common import JobOwners
 from dags.web_preaggregated_utils import (
@@ -27,8 +27,6 @@ from posthog.settings.object_storage import (
     OBJECT_STORAGE_BUCKET,
     OBJECT_STORAGE_ENDPOINT,
     OBJECT_STORAGE_PREAGGREGATED_WEB_ANALYTICS_FOLDER,
-    OBJECT_STORAGE_ACCESS_KEY_ID,
-    OBJECT_STORAGE_SECRET_ACCESS_KEY,
 )
 
 
@@ -261,261 +259,3 @@ def web_pre_aggregate_daily_schedule(context: dagster.ScheduleEvaluationContext)
             }
         },
     )
-
-
-# Asset checks for daily tables
-@asset_check(
-    asset=web_bounces_daily,
-    name="bounces_daily_has_data",
-    description="Check if web_bounces_daily table has data",
-)
-def bounces_daily_has_data() -> AssetCheckResult:
-    """
-    Check if the web_bounces_daily table has data.
-    """
-    try:
-        result = sync_execute("SELECT COUNT(*) FROM web_bounces_daily LIMIT 1")
-        row_count = result[0][0] if result and result[0] else 0
-
-        passed = row_count > 0
-
-        return AssetCheckResult(
-            passed=passed,
-            description=f"Table has {row_count} rows" if passed else "Table is empty",
-            metadata={"row_count": MetadataValue.int(row_count), "table_name": MetadataValue.text("web_bounces_daily")},
-        )
-    except Exception as e:
-        return AssetCheckResult(
-            passed=False, description=f"Error checking table: {str(e)}", metadata={"error": MetadataValue.text(str(e))}
-        )
-
-
-@asset_check(
-    asset=web_stats_daily,
-    name="stats_daily_has_data",
-    description="Check if web_stats_daily table has data",
-)
-def stats_daily_has_data() -> AssetCheckResult:
-    """
-    Check if the web_stats_daily table has data.
-    """
-    try:
-        result = sync_execute("SELECT COUNT(*) FROM web_stats_daily LIMIT 1")
-        row_count = result[0][0] if result and result[0] else 0
-
-        passed = row_count > 0
-
-        return AssetCheckResult(
-            passed=passed,
-            description=f"Table has {row_count} rows" if passed else "Table is empty",
-            metadata={"row_count": MetadataValue.int(row_count), "table_name": MetadataValue.text("web_stats_daily")},
-        )
-    except Exception as e:
-        return AssetCheckResult(
-            passed=False, description=f"Error checking table: {str(e)}", metadata={"error": MetadataValue.text(str(e))}
-        )
-
-
-# Asset checks for export files - verify chdb can query them and they have data
-@asset_check(
-    asset=web_stats_daily_export,
-    name="stats_export_chdb_queryable",
-    description="Check if stats export file can be queried with chdb and has data",
-)
-def stats_export_chdb_queryable() -> AssetCheckResult:
-    """
-    Check if the web_stats_daily export file can be queried with chdb and contains data.
-    """
-    try:
-        import chdb
-        import structlog
-
-        logger = structlog.get_logger(__name__)
-
-        # Get the export path - follow same pattern as export_web_analytics_data
-        if DEBUG:
-            # For local development, use Minio with s3() function for proper authentication
-            s3_endpoint = "http://objectstorage:19000"
-            bucket = "posthog"
-            key = f"{OBJECT_STORAGE_PREAGGREGATED_WEB_ANALYTICS_FOLDER}/web_stats_daily_export.native"
-        else:
-            # For production, use S3 URL (same pattern as export method)
-            s3_endpoint = "https://s3.amazonaws.com"
-            bucket = DAGSTER_DATA_EXPORT_S3_BUCKET
-            key = f"{OBJECT_STORAGE_PREAGGREGATED_WEB_ANALYTICS_FOLDER}/web_stats_daily_export.native"
-
-        try:
-            # Try to query the export file with chdb using s3() function for proper AWS auth
-            s3_query = f"SELECT COUNT(*) as row_count FROM s3('{s3_endpoint}/{bucket}/{key}', '{OBJECT_STORAGE_ACCESS_KEY_ID}', '{OBJECT_STORAGE_SECRET_ACCESS_KEY}', 'Native')"
-            result = chdb.query(s3_query)
-            row_count = int(result.data().strip()) if result.data().strip().isdigit() else 0
-
-            passed = row_count > 0
-            env_type = "Minio" if DEBUG else "S3"
-
-            export_path = f"{s3_endpoint}/{bucket}/{key}"
-            logger.info(
-                "stats_export_chdb_check", export_path=export_path, row_count=row_count, queryable=True, env=env_type
-            )
-
-            return AssetCheckResult(
-                passed=passed,
-                description=f"Export file queryable with chdb via {env_type}, contains {row_count} rows"
-                if passed
-                else f"Export file queryable via {env_type} but empty",
-                metadata={
-                    "row_count": MetadataValue.int(row_count),
-                    "export_path": MetadataValue.text(export_path),
-                    "chdb_queryable": MetadataValue.bool(True),
-                    "export_type": MetadataValue.text("web_stats_daily"),
-                    "environment": MetadataValue.text(env_type),
-                },
-            )
-
-        except (FileNotFoundError, Exception) as e:
-            # File doesn't exist yet or has format issues - this is ok for a check
-            error_msg = str(e)
-            env_type = "Minio" if DEBUG else "S3"
-
-            if "Cannot stat file" in error_msg or "CANNOT_STAT" in error_msg:
-                status = "file_not_found"
-                description = f"Export file not found on {env_type} - may not have been created yet"
-            elif "table structure cannot be extracted" in error_msg.lower():
-                status = "format_issue"
-                description = f"Export file exists on {env_type} but chdb cannot determine structure - this is expected for Native format"
-            else:
-                status = "other_error"
-                description = f"Export file check failed on {env_type}: {error_msg}"
-
-            export_path = f"{s3_endpoint}/{bucket}/{key}"
-            logger.info(
-                "stats_export_chdb_check", export_path=export_path, status=status, error=error_msg[:100], env=env_type
-            )
-
-            return AssetCheckResult(
-                passed=False,
-                description=description,
-                metadata={
-                    "export_path": MetadataValue.text(export_path),
-                    "error": MetadataValue.text(error_msg[:200]),
-                    "export_type": MetadataValue.text("web_stats_daily"),
-                    "status": MetadataValue.text(status),
-                    "environment": MetadataValue.text(env_type),
-                },
-            )
-
-    except ImportError:
-        return AssetCheckResult(
-            passed=False,
-            description="chdb not available - cannot verify export queryability",
-            metadata={"error": MetadataValue.text("chdb import failed")},
-        )
-    except Exception as e:
-        logger.exception("stats_export_chdb_check_error", error=str(e))
-        return AssetCheckResult(
-            passed=False,
-            description=f"Error checking export with chdb: {str(e)}",
-            metadata={"error": MetadataValue.text(str(e))},
-        )
-
-
-@asset_check(
-    asset=web_bounces_daily_export,
-    name="bounces_export_chdb_queryable",
-    description="Check if bounces export file can be queried with chdb and has data",
-)
-def bounces_export_chdb_queryable() -> AssetCheckResult:
-    """
-    Check if the web_bounces_daily export file can be queried with chdb and contains data.
-    """
-    try:
-        import chdb
-        import structlog
-
-        logger = structlog.get_logger(__name__)
-
-        # Get the export path - follow same pattern as export_web_analytics_data
-        if DEBUG:
-            # For local development, use Minio with s3() function for proper authentication
-            s3_endpoint = "http://objectstorage:19000"
-            bucket = "posthog"
-            key = f"{OBJECT_STORAGE_PREAGGREGATED_WEB_ANALYTICS_FOLDER}/web_bounces_daily_export.native"
-        else:
-            # For production, use S3 URL (same pattern as export method)
-            s3_endpoint = "https://s3.amazonaws.com"
-            bucket = DAGSTER_DATA_EXPORT_S3_BUCKET
-            key = f"{OBJECT_STORAGE_PREAGGREGATED_WEB_ANALYTICS_FOLDER}/web_bounces_daily_export.native"
-
-        try:
-            # Try to query the export file with chdb using s3() function for proper AWS auth
-            s3_query = f"SELECT COUNT(*) as row_count FROM s3('{s3_endpoint}/{bucket}/{key}', '{OBJECT_STORAGE_ACCESS_KEY_ID}', '{OBJECT_STORAGE_SECRET_ACCESS_KEY}', 'Native')"
-            result = chdb.query(s3_query)
-            row_count = int(result.data().strip()) if result.data().strip().isdigit() else 0
-
-            passed = row_count > 0
-            env_type = "Minio" if DEBUG else "S3"
-
-            export_path = f"{s3_endpoint}/{bucket}/{key}"
-            logger.info(
-                "bounces_export_chdb_check", export_path=export_path, row_count=row_count, queryable=True, env=env_type
-            )
-
-            return AssetCheckResult(
-                passed=passed,
-                description=f"Export file queryable with chdb via {env_type}, contains {row_count} rows"
-                if passed
-                else f"Export file queryable via {env_type} but empty",
-                metadata={
-                    "row_count": MetadataValue.int(row_count),
-                    "export_path": MetadataValue.text(export_path),
-                    "chdb_queryable": MetadataValue.bool(True),
-                    "export_type": MetadataValue.text("web_bounces_daily"),
-                    "environment": MetadataValue.text(env_type),
-                },
-            )
-
-        except (FileNotFoundError, Exception) as e:
-            # File doesn't exist yet or has format issues - this is ok for a check
-            error_msg = str(e)
-            env_type = "Minio" if DEBUG else "S3"
-
-            if "Cannot stat file" in error_msg or "CANNOT_STAT" in error_msg:
-                status = "file_not_found"
-                description = f"Export file not found on {env_type} - may not have been created yet"
-            elif "table structure cannot be extracted" in error_msg.lower():
-                status = "format_issue"
-                description = f"Export file exists on {env_type} but chdb cannot determine structure - this is expected for Native format"
-            else:
-                status = "other_error"
-                description = f"Export file check failed on {env_type}: {error_msg}"
-
-            export_path = f"{s3_endpoint}/{bucket}/{key}"
-            logger.info(
-                "bounces_export_chdb_check", export_path=export_path, status=status, error=error_msg[:100], env=env_type
-            )
-
-            return AssetCheckResult(
-                passed=False,
-                description=description,
-                metadata={
-                    "export_path": MetadataValue.text(export_path),
-                    "error": MetadataValue.text(error_msg[:200]),
-                    "export_type": MetadataValue.text("web_bounces_daily"),
-                    "status": MetadataValue.text(status),
-                    "environment": MetadataValue.text(env_type),
-                },
-            )
-
-    except ImportError:
-        return AssetCheckResult(
-            passed=False,
-            description="chdb not available - cannot verify export queryability",
-            metadata={"error": MetadataValue.text("chdb import failed")},
-        )
-    except Exception as e:
-        logger.exception("bounces_export_chdb_check_error", error=str(e))
-        return AssetCheckResult(
-            passed=False,
-            description=f"Error checking export with chdb: {str(e)}",
-            metadata={"error": MetadataValue.text(str(e))},
-        )
